feat: add delete-user use case

+++ src/app/use-cases/delete-user.ts
@@ -4,7 +4,6 @@ import { UserRepository } from "../repositories/user-repository"
 
 interface IDeleteUserUseCaseRequest {
   userId: string
-  email: string
 }
 
 
@@ -18,7 +17,11 @@ export class DeleteUserUseCase{
   }
   async execute({userId}: IDeleteUserUseCaseRequest): Promise<void>{
  
+    const userSuperAdmin = await this.userRepository.userIsSuperAdmin(userId)
 
+    if (userSuperAdmin){
+      throw new Error('Not Authorized')
+    }
 
     await this.userRepository.delete(userId)
